Moving validation methods into their own module
